SpinGlassExhaustive

Index

SpinGlassExhaustive.energy

SpinGlassExhaustive.energy_qubo

SpinGlassExhaustive.exhaustive_search

SpinGlassExhaustive.exhaustive_search_bucket

SpinGlassExhaustive.generate_random_graph

SpinGlassExhaustive.get_energy_offset

SpinGlassExhaustive.graph_to_dict

SpinGlassExhaustive.graph_to_qubo

SpinGlassExhaustive.kernel

SpinGlassExhaustive.kernel_bucket

SpinGlassExhaustive.kernel_part

SpinGlassExhaustive.kernel_qubo

SpinGlassExhaustive.max_chunk_size

SpinGlassExhaustive.partial_exhaustive_search

SpinGlassExhaustive.energy — Method

energy(state_code, graph)

state_code: state code for which the energy.

graph: graph of the ising model.

Returns the state energy.

SpinGlassExhaustive.energy_qubo — Method

energy_qubo(state_code, graph)

state_code: state code for which the energy expressed in qubo is to be calculated.

graph: graph of the ising model.

Returns the state energy expressed as QUBO.

SpinGlassExhaustive.exhaustive_search — Method

exhaustive_search(ig)

ig::IsingGraph: graph of ising model represented by IsingGraph structure.

Returns energies and states for provided graph by brute-forece alorithm based on GPU.

SpinGlassExhaustive.exhaustive_search_bucket — Function

exhaustive_search_bucket(ig)
exhaustive_search_bucket(ig, how_many)

ig::IsingGraph: graph of ising model represented by IsingGraph structure.

Returns energies and states for provided graph by brute-forece alorithm supported by bucket selection based on GPU.

SpinGlassExhaustive.generate_random_graph — Method

generate_random_graph(d)

d::Int: size of random graph.

Returns random array of size d.

SpinGlassExhaustive.get_energy_offset — Method

get_energy_offset(graph)

graph: graph of the ising model.

Returns offest between Ising model graph and its QUBO representation.

SpinGlassExhaustive.graph_to_dict — Method

graph_to_dict(graph)

graph: graph of the ising model.

Converts Ising model graph to Dict.

SpinGlassExhaustive.graph_to_qubo — Method

graph_to_qubo(graph)

graph: graph of the ising model.

Converts Ising model graph to QUBO representation.

SpinGlassExhaustive.kernel — Method

kernel(graph, energies)

graph: graph of the ising model.

energies: array filled with zeros. Each array index represents the state of the system.

Returns energies for every state.

SpinGlassExhaustive.kernel_bucket — Method

kernel_bucket(graph, energies, idx)

graph: graph of the ising model.

energies: array filled with zeros. Each array index represents the state of the system.

idx: list for collecting partial energy results.

Returns energies for given indexes.

SpinGlassExhaustive.kernel_part — Method

kernel_part(graph, energies, part_lst, part_st)

graph: graph of the ising model.

energies: array filled with zeros. Each array index represents the state of the system.

part_lst: list for collecting partial energy results.

part_st: list for collecting partial state results.

Returns energies for every state.

SpinGlassExhaustive.kernel_qubo — Method

kernel_qubo(graph, energies)

graph: graph of the ising model.

energies: array filled with zeros. Each array index represents the state of the system.

Returns energies energy expressed as QUBO for every state.

SpinGlassExhaustive.max_chunk_size — Method

max_chunk_size()

Returns the maximum chunk size for the algorithm supported by bucket selection.

SpinGlassExhaustive.partial_exhaustive_search — Method

partial_exhaustive_search(ig)

ig::IsingGraph: graph of ising model represented by IsingGraph structure.

Returns energies and states for provided graph by brute-forece alorithm supported by partial selection based on GPU.